requirements.txt
autopep8==1.4.4
azure-cognitiveservices-speech==1.8.0
azure-cognitiveservices-vision-face==0.4.0
azure-common==1.1.23
beautifulsoup4==4.8.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
msrest==0.6.10
numpy==1.17.3
oauthlib==3.1.0
opencv-python==4.1.1.26
Pillow==6.2.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyAudio==0.2.11
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
twilio==6.33.0
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
